+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 2.4.1 [unreleased]
 
+- Better exception when trying to access results for failed multi-search query
 - Added support for `faraday_middleware-aws-sigv4`
 - Added `credentials` option to `aws_credentials`